+++ arcore-location-extension/src/main/kotlin/org/neidhardt/arlocation/LocationArMarker.kt
@@ -10,10 +10,10 @@ class LocationArMarker(
 	var anchorNode: LocationArNode? = null
 
 	/**
-	 * [onRender] is called on each frame an can be used to update marker
+	 * [onUpdate] is called on each frame an can be used to update marker
 	 * properties before it is being rendered.
 	 */
-	var onRender: NodeRenderEventListener? = null
+	var onUpdate: NodeUpdateEventListener? = null
 
 	/**
 	 * [height] is the altitude of the object, based on camera height.