+++ arcore-location-extension/src/main/kotlin/org/neidhardt/arlocation/NodeRenderEventListener.kt
@@ -1,5 +1,0 @@
-package org.neidhardt.arlocation
-
-interface NodeRenderEventListener {
-	fun onRender(node: LocationArNode)
-}