+++ arcore-location-extension/src/main/kotlin/org/neidhardt/arlocation/NodeUpdateEventListener.kt
@@ -1,0 +1,10 @@
+package org.neidhardt.arlocation
+
+import com.google.ar.sceneform.FrameTime
+
+/**
+ * [NodeUpdateEventListener] is called each frame before the node is rendered (onUpdate callback).
+ */
+interface NodeUpdateEventListener {
+	fun onUpdate(frameTime: FrameTime, locationNode: LocationArNode)
+}